// src/lib/types/LawAdoptionDataPoint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type LawAdoptionDataPoint = { turn: number, law_count: number, 
/**
 * The name of the law adopted at this point (None for synthetic start/end points)
 */
law_name: string | null, };