// src/lib/types/StoryEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface StoryEvent {
  event_id: number;
  event_type: string;
  player_name: string;
  occurred_turn: number;
  primary_character_name: string | null;
  city_name: string | null;
}